% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds_connect.R
\name{wrds_connect}
\alias{wrds_connect}
\title{Open connection to WRDS}
\usage{
wrds_connect(username)
}
\arguments{
\item{username}{string of the username you have with WRDS.}
}
\description{
Before being able to open the connection, you need to follow the steps
\href{https://wrds-www.wharton.upenn.edu/pages/support/programming-wrds/programming-r/r-from-your-computer/}{here} to set up your \code{.pgpass} file.
This function should be run before all other wrds. functions, to open the database connection.
}
\examples{
wrds_connect( "preggian")

}
