% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds_inspect.R
\name{wrds_inspect}
\alias{wrds_inspect}
\title{Observe Tables structure in WRDS}
\usage{
wrds_inspect(wrds)
}
\arguments{
\item{wrds}{connection created with the wrds.connect() function.}
}
\value{
a data.frame containing the WRDS database tables structure
}
\description{
Remember to change the username in the dbConnect call below. Before being able to open the connection, you need to follow the steps
\href{https://wrds-www.wharton.upenn.edu/pages/support/programming-wrds/programming-r/r-from-your-computer/}{here} to set up your \code{.pgpass} file.
}
