% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds_read.R
\name{wrds_read}
\alias{wrds_read}
\title{Take variables from table in WRDS}
\usage{
wrds_read(query, wrds, numrows = -1, data.table = TRUE)
}
\arguments{
\item{query}{string including the SQL query syntax for WRDS (see their webpage)}

\item{wrds}{connection created with the \code{wrds_connect()} function.}

\item{numrows}{number of observations you want to get, default is all the available ones.}

\item{data.table}{set to FALSE if you want a normal data.frame.}
}
\value{
a data.table or .frame containing the WRDS data requested
}
\description{
Before being able to open the connection, you need to follow the steps
\href{https://wrds-www.wharton.upenn.edu/pages/support/programming-wrds/programming-r/r-from-your-computer/}{here} to set up your \code{.pgpass} file.
}
\examples{
data <- wrds_read( "select mgrname from tfn.s34 where fdate between '2013-01-02' and '2015-03-02' ", numrows=100)

}
