% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds_tables.R
\name{wrds_tables}
\alias{wrds_tables}
\title{Observe Data Tables in chosen WRDS folder}
\usage{
wrds_tables(folder, wrds)
}
\arguments{
\item{folder}{folder in WRDS that you want to inspect, find it with \code{wrds.inspect()}.}

\item{wrds}{connection created with the wrds.connect() function.}
}
\value{
a data.frame containing the WRDS database tables structure
}
\description{
Remember to change the username in the dbConnect call below. Before being able to open the connection, you need to follow the steps
\href{https://wrds-www.wharton.upenn.edu/pages/support/programming-wrds/programming-r/r-from-your-computer/}{here} to set up your \code{.pgpass} file.
Usually this function is used after wrds.inspect() to check what data tables are included in a WRDS product.
For instance if you want to know the tables under tfn for the Thomson Reuters data.
}
